// src/app/services/swagger/model/jWTRes.ts
/**
 * BoF REST API
 * description
 *
 * OpenAPI spec version: 0.0.1
 * Contact: bla@bla.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface JWTRes { 
    /**
     * The access token to authenticate every request against the backend.
     */
    accessToken: string;
    /**
     * The user's ID.
     */
    idUser: number;
    /**
     * The access token to authenticate every request against the backend.
     */
    refreshToken: string;
    /**
     * The user's role.
     */
    role: JWTRes.RoleEnum;
    /**
     * The user's name.
     */
    username: string;
}
export namespace JWTRes {
    export type RoleEnum = 'USER';
    export const RoleEnum = {
        USER: 'USER' as RoleEnum
    };
}